+++ tests/Unit/Listeners/FlushAuthenticationStateListenerTest.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace Spiral\RoadRunnerLaravel\Tests\Unit\Listeners;
 
 use Mockery as m;
-use Illuminate\Support\Str;
-use Illuminate\Cache\CacheManager;
 use Spiral\RoadRunnerLaravel\Events\Contracts\WithApplication;
 use Spiral\RoadRunnerLaravel\Listeners\FlushAuthenticationStateListener;